pyproject.toml
[project]
name = "daily-eff-oldway"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "google-auth>=2.43.0",
    "gspread>=6.2.1",
    "gspread-dataframe>=4.0.0",
    "oauth2client>=4.1.3",
    "pandas>=2.3.3",
    "pyodbc>=5.3.0",
]
